+++ .examples/tic-tac-toe/src/app/on-winner.ts
@@ -1,0 +1,15 @@
+import { getTheseusLogger } from "theseus-js";
+import { GameShip } from "../game-ship/game-ship";
+import { MarkType } from "../game-ship/state/GameState";
+
+const log = getTheseusLogger("OnWinner");
+
+export const onWinner = () => 
+{
+	const foundTriple = GameShip.refine.Outcome.checkForTriples();
+	if (foundTriple) 
+	{
+		const { markType, tripleTypePlainEnglish } = GameShip.refine.Outcome.getTripleType(foundTriple);
+		log.major(`We have a winner! Three ${markType}s ${tripleTypePlainEnglish}!`);
+	}
+};
